requirements.txt
streamlit
openai
pinecone-client
langchain
Langchain-community
Time